[#508] Extract SleekDB related-criteria workflow into statement trait
